Moved device storage tests
